pyproject.toml
[tool.poetry]
name ="lab4"
version = "0.1.0"
description = "Berkeley MIDS W255 Lab4"
authors = ["jooyeonhahm <jooyeonhahm@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
fastapi = "^0.95.2"
uvicorn = "^0.22.0"
requests = "^2.31"
joblib = "^1.2.0"
scipy = "^1.10"
scikit-learn = "1.2.2"
fastapi-cache2 = "0.1.9"
redis = "^4.6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3"
black = "^23.3"
isort = "^5.12"
httpx = "^0.24"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
